app/requirements.txt
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
pyjwt
